+++ LayerKit.embeddedframework/LayerKit.framework/Versions/A/Headers/LYRMessage.h
@@ -129,7 +129,7 @@ extern NSString * _Nonnull const LYRMessageOptionsPushNotificationConfigurationK
 
 /**
  @abstract Deletes a message in the specified mode.
- @param mode The deletion mode, specifying how the message is to be deleted (i.e. locally or synchronized across participants).
+ @param mode The deletion mode, specifying how the message is to be deleted (i.e. for only the currently authenticated user's devices or synchronized across participants).
  @param error A pointer to an error that upon failure is set to an error object describing why the deletion failed.
  @return A Boolean value indicating if the request to delete the message was submitted for synchronization.
  @raises NSInvalidArgumentException Raised if `message` is `nil`.